+++ scripts/create-sd-card-image.sh
@@ -53,6 +53,13 @@ sudo mkdir -p /mnt/data/Scripts
 wget -c https://raw.githubusercontent.com/MiSTer-devel/Scripts_MiSTer/master/other_authors/wifi.sh
 sudo cp wifi.sh /mnt/data/Scripts
 
+# Bundle new downloader script with Mr. Fusion
+# TODO(m): Remove this when the upstream MiSTer release includes the new downloader script
+# by default.
+# See https://github.com/MiSTer-devel/Main_MiSTer/issues/483
+wget -c https://raw.githubusercontent.com/MiSTer-devel/Downloader_MiSTer/main/downloader.sh
+sudo cp downloader.sh /mnt/data/Scripts
+
 # Support custom MiSTer config
 sudo mkdir -p /mnt/data/config